refactor(Masthead): Create hook for menu items
